refactor(tests): colocate unit test files next to source modules
