+++ src/features/positions/stores/fetcher.test.ts
@@ -1,11 +1,11 @@
-import { fetchPositions } from '../../stores/fetcher';
-import { FIXTURE_LIST_POSITIONS_SUCCESS, FIXTURE_PARAMS } from '../../__fixtures__/ListPositions';
+import { fetchPositions } from './fetcher';
+import { FIXTURE_LIST_POSITIONS_SUCCESS, FIXTURE_PARAMS } from '../__fixtures__/ListPositions';
 
 jest.mock('@/config', () => ({
   getExperimentalFlag: jest.fn(() => false),
   DEFI_POSITIONS_THRESHOLD_FILTER: 'defiPositionsThresholdFilter',
 }));
-jest.mock('../../stores/fetcher', () => ({
+jest.mock('./fetcher', () => ({
   fetchPositions: jest.fn(),
 }));
