+++ src/screens/rewards/helpers/formatTokenDisplayValue.test.ts
