+++ src/screens/rewards/helpers/getGradientColorsForRank.test.ts
