+++ src/screens/transaction-details/helpers/shortenTxHashString.test.ts
